fix: BBOX兼容ARM平台
1、在核心框架中使用基础数据类型。
2、尽量使用用户层的函数接口。

Close #I422VS

+++ blackbox/blackbox_detector.c
@@ -24,10 +24,10 @@ int UploadEventByFile(const char *filePath)
     return 0;
 }
 
-int UploadEventByStream(const char *buf, size_t bufSize)
+int UploadEventByStream(const char *buf, unsigned int bufSize)
 {
     if (buf == NULL || bufSize == 0) {
-        BBOX_PRINT_ERR("buf: %p, bufSize: %u\n", buf, (UINT32)bufSize);
+        BBOX_PRINT_ERR("buf: %p, bufSize: %u\n", buf, bufSize);
         return -1;
     }
     return 0;
